backend/requirements.txt
Django>=4.0
djangorestframework>=3.13
djongo>=1.3.6
django-cors-headers>=3.11
pymongo==3.12.1
pytz
sqlparse

